+++ CHANGELOG.md
@@ -15,8 +15,13 @@
     * The `Document` type no longer implements these traits: `Eq`,
       `Hash`, `Ord`, and `PartialOrd`.
 * The following types now have at least one private field and can no
-  longer be directly constructed by applications: `Database`, `Design`,
-  `ErrorResponse`, and `ViewFunction`.
+  longer be directly constructed by applications:
+    * `Database`
+    * `Design`
+    * `ErrorResponse`
+    * `ViewFunction`
+    * `ViewResult`
+    * `ViewRow`
 
 ### Deprecations